js/sass scaffolding changes
