+++ frontend/resources/assets/js/components/ui/global/headroom-init.js
@@ -1,5 +1,3 @@
-import Headroom from 'headroom.js';
-
 function init(){
   // grab an element
   var header = document.querySelector("header");
