+++ frontend/resources/assets/js/components/ui/global/load.js
@@ -20,23 +20,8 @@ function _pagePreload() {
   });
 }
 
-function _scrollTriggerAnimation(){
-  // On-scroll and on-view trigger animation
-  if( $('[data-animate]').length ) {
-    $('[data-animate]').each(function () {
-      var inview = new Waypoint.Inview({
-        element: this,
-        enter: function(direction) {
-          $(this.element).addClass('show');
-        }
-      });
-    });
-  }
-}
-
 function init() {
   // _pagePreload();
-  _scrollTriggerAnimation();
 }
 
 export default init;
