+++ frontend/resources/assets/js/components/ui/global/slick.js
@@ -1,5 +1,3 @@
-import slick from 'slick-carousel';
-
 function init(){
   // Carousel 1
   // ---
